chore(*): fix more tests
