+++ src/commands/create/spec/create.spec.ts
@@ -3,7 +3,7 @@ import path from 'path'
 import { verifyFolderStructure } from '../../../utils/test'
 import { createFolder, deleteFolder, fileExists } from '../../../utils/file'
 import { asyncForEach, generateTimestamp } from '../../../utils/utils'
-import { getFolders } from '../../../utils/config-utils'
+import { getFolders } from '../../../utils/config'
 import fileStructureMinimalObj from './files-minimal-app.json'
 import fileStructureReactObj from './files-react-app.json'
 import fileStructureAngularrObj from './files-angular-app.json'
