+++ src/commands/job/spec/job.server.spec.ts
@@ -17,7 +17,7 @@ import { ServerType, Target } from '@sasjs/utils/types'
 import {
   removeFromGlobalConfig,
   saveToGlobalConfig
-} from '../../../utils/config-utils'
+} from '../../../utils/config'
 import { Command } from '../../../utils/command'
 import { createTestApp, removeTestApp } from '../../../utils/test'
