+++ src/utils/test.ts
@@ -2,7 +2,7 @@ import path from 'path'
 import dotenv from 'dotenv'
 import { deleteFolder, fileExists, folderExists } from './file'
 import { ServerType, Target } from '@sasjs/utils/types'
-import { saveToGlobalConfig } from './config-utils'
+import { saveToGlobalConfig } from './config'
 import fileStructureDBObj from './fileStructures/files-db.json'
 import fileStructureCompileObj from './fileStructures/files-compiled.json'
 import fileStructureBuildObj from './fileStructures/files-built.json'
