--- test/sample-sas.js
@@ -1,36 +1,0 @@
-export const sampleSasProgram = `/**
-  @file mv_createfolder.sas
-  @brief Creates a viya folder if that foloder does not already exist
-  @details Expects oauth token in a global macro variable (default
-    ACCESS_TOKEN).
-
-    options mprint;
-    %mv_createfolder(path=/Public)
-
-
-  @param path= The full path of the folder to be created
-  @param access_token_var= The global macro variable to contain the access token
-  @param grant_type= valid values are "password" or "authorization_code" (unquoted).
-    The default is authorization_code.
-
-
-  @version VIYA V.03.04
-  @author Allan Bowe
-  @source https://github.com/sasjs/core
-
-  <h4> Dependencies </h4>
-  @li mf_abort.sas
-  @li mf_getuniquefileref.sas
-  @li mf_getuniquelibref.sas
-  @li mf_isblank.sas
-
-**/
-
-/* This is a comment
-that spans
-multiple lines */
-
-%macro mv_createfolder(path=
-  ,access_token_var=ACCESS_TOKEN
-  ,grant_type=authorization_code
-);`
